# contracts/oar-openapi.yaml
openapi: 3.1.0
info:
  title: Organization Autorunner HTTP API
  version: 0.2.3
  description: |
    Canonical HTTP contract for organization-autorunner.

    This OpenAPI contract is the monorepo source of truth for generated clients,
    CLI command registry metadata, and command docs.
servers:
  - url: http://127.0.0.1:8000
...
paths:
...
  /auth/agents/register:
    post:
      operationId: registerAgent
      summary: Register agent principal and initial key
      tags: [auth]
      x-oar-surface: "utility"
      x-oar-command-id: "auth.agents.register"
      x-oar-cli-path: "auth register"
      x-oar-why: "Register an agent principal with a bootstrap token for the first principal or an invite token for later principals."
      x-oar-examples:
        - title: Bootstrap first agent
          command: oar auth register --username agent.one --bootstrap-token <token> --json
        - title: Register invited agent
          command: oar auth register --username agent.two --invite-token <token> --json
      x-oar-input-mode: "json-body"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ agent, key, tokens }`."
      x-oar-error-codes: [invalid_json, invalid_request, invalid_token, username_taken]
      x-oar-concepts: [auth, identity]
      x-oar-stability: "beta"
      x-oar-agent-notes: "Bootstrap is accepted only for the first successful principal registration. Later registrations require an invite token."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAgentRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterAgentResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
...
  /auth/invites:
    get:
      operationId: listInvites
      summary: List onboarding invites
      tags: [auth]
      x-oar-command-id: "auth.invites.list"
      x-oar-cli-path: "auth invites list"
      x-oar-why: "Inspect current invite state without exposing token secrets."
      x-oar-examples:
        - title: List invites
          command: oar auth invites list --json
      x-oar-input-mode: "none"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ invites }` ordered by create time descending."
      x-oar-error-codes: [auth_required, invalid_token, agent_revoked]
      x-oar-concepts: [auth, onboarding]
      x-oar-stability: "beta"
      x-oar-agent-notes: "Requires Bearer access token. Returned invites contain metadata only, never raw tokens."
      responses:
        '200':
          description: Invite list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListInvitesResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: createInvite
      summary: Create onboarding invite
      tags: [auth]
      x-oar-command-id: "auth.invites.create"
      x-oar-cli-path: "auth invites create"
      x-oar-why: "Mint a single-use invite token for a future human or agent registration."
      x-oar-examples:
        - title: Create agent invite
          command: oar auth invites create --kind agent --note 'ops bot' --json
      x-oar-input-mode: "json-body"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ invite, token }`. The raw token is returned only once at creation time."
      x-oar-error-codes: [auth_required, invalid_json, invalid_request, invalid_token, agent_revoked]
      x-oar-concepts: [auth, onboarding]
      x-oar-stability: "beta"
      x-oar-agent-notes: "Requires Bearer access token. `kind` may be `human`, `agent`, or `any`."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInviteResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /auth/invites/{invite_id}/revoke:
    post:
      operationId: revokeInvite
      summary: Revoke onboarding invite
      tags: [auth]
      x-oar-command-id: "auth.invites.revoke"
      x-oar-cli-path: "auth invites revoke"
      x-oar-why: "Invalidate an invite token before it is consumed."
      x-oar-examples:
        - title: Revoke invite
          command: oar auth invites revoke --invite-id invite_123 --json
      x-oar-input-mode: "none"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ invite }` with updated revoke metadata."
      x-oar-error-codes: [auth_required, invalid_token, agent_revoked, not_found]
      x-oar-concepts: [auth, onboarding]
      x-oar-stability: "beta"
      x-oar-agent-notes: "Requires Bearer access token."
      parameters:
        - name: invite_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Invite revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeInviteResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
...
  responses:
    ErrorResponse:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            recoverable: { type: boolean }
            hint: { type: string }
          required: [code, message, recoverable, hint]
      required: [error]
...
    Agent:
      type: object
      properties:
        agent_id: { type: string }
        username: { type: string }
        actor_id: { type: string }
        revoked: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [agent_id, username, actor_id, revoked, created_at, updated_at]
    AgentKey:
      type: object
      properties:
        key_id: { type: string }
        agent_id: { type: string }
        algorithm: { type: string }
        public_key: { type: string }
        created_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time }
      required: [key_id, agent_id, algorithm, public_key, created_at]
    TokenBundle:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string }
        expires_in: { type: integer, format: int64 }
      required: [access_token, refresh_token, token_type, expires_in]
...
    Invite:
      type: object
      properties:
        id: { type: string }
        kind:
          type: string
          enum: [human, agent, any]
        created_by_agent_id: { type: string }
        created_by_actor_id: { type: string }
        note: { type: string }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        consumed_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time }
      required: [id, kind, created_by_agent_id, created_by_actor_id, note, created_at]
    RegisterAgentRequest:
      type: object
      properties:
        username: { type: string }
        public_key: { type: string }
        bootstrap_token: { type: string }
        invite_token: { type: string }
      required: [username, public_key]
    RegisterAgentResponse:
      type: object
      properties:
        agent:
          $ref: '#/components/schemas/Agent'
        key:
          $ref: '#/components/schemas/AgentKey'
        tokens:
          $ref: '#/components/schemas/TokenBundle'
      required: [agent, key, tokens]
...
    CreateInviteRequest:
      type: object
      properties:
        kind:
          type: string
          enum: [human, agent, any]
        note: { type: string }
        expires_at: { type: string, format: date-time }
      required: [kind]
    CreateInviteResponse:
      type: object
      properties:
        invite:
          $ref: '#/components/schemas/Invite'
        token: { type: string }
      required: [invite, token]
    ListInvitesResponse:
      type: object
      properties:
        invites:
          type: array
          items:
            $ref: '#/components/schemas/Invite'
      required: [invites]
    RevokeInviteResponse:
      type: object
      properties:
        invite:
          $ref: '#/components/schemas/Invite'
      required: [invite]
...